chore: Reorganized the files chore: documented General Math File
