+++ Sources/SwiftDithering/Dithers/Bayer.swift
@@ -1,0 +1,70 @@
+//
+//  File.swift
+//  
+//
+//  Created by Lugalu on 30/07/23.
+//
+
+import Foundation
+
+/**
+ This enum contains all the information regarding Bayer matrices from size to matrix values.
+ */
+public enum BayerSizes: Int{
+    case bayer2x2 = 2
+    case bayer4x4 = 4
+    case bayer8x8 = 8
+    
+    
+    func getBayerMatrix() -> [[UInt8]]{
+        switch self{
+        case .bayer2x2:
+            return [
+                [0, 2],
+                [3, 1]
+            ]
+        case .bayer4x4:
+            return [
+                [0, 8, 2, 10],
+                [12, 4, 14, 6],
+                [3, 11, 1, 9],
+                [15, 7, 13, 5]
+            ]
+        case .bayer8x8:
+            return [
+                [ 0, 32,  8, 40,  2, 34, 10, 42],
+                [48, 16, 56, 24, 50, 18, 58, 26],
+                [12, 44,  4, 36, 14, 46,  6, 38],
+                [60, 28, 52, 20, 62, 30, 54, 22],
+                [ 3, 35, 11, 43,  1, 33,  9, 41],
+                [51, 19, 59, 27, 49, 17, 57, 25],
+                [15, 47,  7, 39, 13, 45,  5, 37],
+                [63, 31, 55, 23, 61, 29, 53, 21]
+            ]
+        }
+    }
+}
+
+/**
+    Loops through all the pixels on screen then calculate and applies the color, all is done via the pointer so no need for returns all parameters are carried over from [apply ordered dither](x-source-tag://applyOrderedDither)
+ */
+internal func bayerDither(_ imageData: inout UnsafeMutablePointer<UInt8>, bayerSize: BayerSizes, width: Int, height: Int, spread: Double, bytesPerPixel: Int){
+    for y in 0..<height{
+        for x in 0..<width{
+            let index = indexCalculator(x: x, y: y, width: width, bytesPerPixel: bytesPerPixel)
+            
+            let bayerFactor = bayerSize.rawValue
+            let bayerMatrixResult = bayerSize.getBayerMatrix()[y % bayerFactor][x % bayerFactor]
+            let bayerValue = Double(bayerMatrixResult) - 0.5
+            
+            let (oldR, oldG, oldB) = getRgbFor(index: index, inData: imageData)
+            
+            let newR = Int(Double(oldR) + spread * bayerValue)
+            let newG = Int(Double(oldG) + spread * bayerValue)
+            let newB = Int(Double(oldB) + spread * bayerValue)
+            
+            assignNewColorsTo(imageData: &imageData, index: index, colors: (newR, newG, newB))
+            
+        }
+    }
+}
