+++ Sources/SwiftDithering/Dithers/ErrorDifusion.swift
@@ -23,8 +23,8 @@ public extension UIImage{
         - Returns: UIImage with the dithering applied
      */
     /// - Tag: applyErrorDifusion
-    func applyErrorDifusion(withType diffusionType: ErrorDifusionTypes, nearestFactor: Int = 2) throws -> UIImage{
-        guard let cgImageTemp = self.cgImage else { throw ImageErrors.failedToRetriveCGImage(localizedDescription: "needed CGImage is not Available")}
+    func applyErrorDifusion(withType diffusionType: ErrorDifusionTypes, nearestFactor: Int = 2) throws -> UIImage {
+        guard let cgImageTemp = self.cgImage else { throw ImageErrors.failedToRetriveCGImage(localizedDescription: "needed CGImage is not Available") }
         
         let cgImage = try convertColorSpaceToRGB(cgImageTemp)
     
@@ -33,16 +33,27 @@ public extension UIImage{
         
         let quantizatorImage = try convertColorSpaceToGrayScale(cgImage)
         let (_,quantitizedImageData, quantitizedBytesPerPixel) = try prepareQuantization(grayScaleImage: quantizatorImage)
+        
+        defer{
+            quantitizedImageData.deallocate()
+        }
+        
         var (imageContext, imageData, bytesPerPixel) = try createContextAndData(cgImage: cgImage, width: width, height: height)
         
         defer {
             imageData.deallocate()
-            quantitizedImageData.deallocate()
         }
 
         switch diffusionType {
         case .floydSteinberg:
-            floydDither(finalImageData: &imageData, quantitizedImageData: quantitizedImageData, width: width, height: height, finalImageBytesPerPixel: bytesPerPixel, quantitizedBytesPerPixel: quantitizedBytesPerPixel, nearestFactor: nearestFactor)
+            floydDither(finalImageData: &imageData,
+                        quantitizedImageData: quantitizedImageData,
+                        width: width,
+                        height: height,
+                        finalImageBytesPerPixel: bytesPerPixel,
+                        quantitizedBytesPerPixel: quantitizedBytesPerPixel,
+                        nearestFactor: nearestFactor)
+            
         case .stucki:
             //TODO: Implement Stucki type.
             return UIImage(systemName: "pencil")!
@@ -58,49 +69,7 @@ public extension UIImage{
         return UIImage(cgImage: outputCGImage)
     }
     
-    /**
-        Loops through all the pixels on screen then calculate and applies the color for multiple pixels, all is done via the pointer so no need for returns all parameters are carried over from [apply Error dither](x-source-tag://applyErrorDifusion) when the type is set to Floyd-Steinberg.
-     */
-    private func floydDither(finalImageData: inout UnsafeMutablePointer<UInt8>, quantitizedImageData: UnsafeMutablePointer<UInt8>, width: Int, height: Int, finalImageBytesPerPixel finalBytesPerPixel: Int, quantitizedBytesPerPixel: Int, nearestFactor: Int){
-        
-        for y in 0..<height{
-            for x in 0..<width{
-                let finalImageIndex = indexCalculator(x: x, y: y, width: width, bytesPerPixel: finalBytesPerPixel)
-                let quantitizedIndex = indexCalculator(x: x, y: y, width: width, bytesPerPixel: quantitizedBytesPerPixel)
-                let quantitizedValue = Int(quantitizedImageData[quantitizedIndex])
-                
-                let oldColor = (finalImageData[finalImageIndex],
-                                finalImageData[finalImageIndex + 1],
-                                finalImageData[finalImageIndex + 2])
-                
-                let newColor = findClosestPallete(oldColor, nearestFactor: nearestFactor)
-                
-                let error = (
-                    r: quantitizedValue - newColor.r,
-                    g: quantitizedValue - newColor.g,
-                    b: quantitizedValue - newColor.b
-                )
-                
-                assignNewColorsTo(imageData: &finalImageData, index: finalImageIndex, colors: newColor)
-                
-                if x + 1 < width{
-                    applyQuantization(&finalImageData, error, x: x + 1, y: y, width: width, bytesPerPixel: finalBytesPerPixel)
-                }
-                if y + 1 < height{
-                    if x - 1 >= 0 {
-                        applyQuantization(&finalImageData, error, x: x - 1, y: y + 1, width: width, bytesPerPixel: finalBytesPerPixel, multiplier: 3)
-                    }
-                    applyQuantization(&finalImageData, error, x: x, y: y + 1, width: width, bytesPerPixel: finalBytesPerPixel, multiplier: 5)
-                    
-                    if x + 1 < width{
-                        applyQuantization(&finalImageData, error, x: x + 1, y: y + 1, width: width, bytesPerPixel: finalBytesPerPixel, multiplier: 1)
-                    }
-                }
-                
-            }
-        }
-        
-    }
+ 
     
     
 }
